python-bot/requirements.txt
# ============================================
# FILE: python-bot/requirements.txt (Production)
# ============================================

# Core dependencies
python-telegram-bot==20.7
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0

# Dependencies dari python-telegram-bot
anyio==4.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
multidict==6.0.4
sniffio==1.3.0
urllib3==2.1.0
yarl==1.9.4

# Timezone
pytz==2023.3
